:whale: Complete Dockerfile :pencil: Update README
I worked on the Dockerfile and got it to a working environment. All
dependencies are installed and working. Running `pipenv run python gen-key.py`
is creating an error; however, not one involving the environment. I have
renamed README.md to README-Original.mb because I have added a new README.md.
This new README documents how to build and use the Dockerfile